Release 0.9.0: neo-vm-rs 0.2.0 migration + Neo N3 v3.10.0 parity fixes
Migrate the workspace to the neo-vm-rs 0.2.0 StackValue API: the compound
variants (Buffer/Array/Struct/Map) now carry a leading, non-serialized u64
reference id and next_stack_item_id() returns u64. Binary/JSON wire formats
and consensus hashes are unchanged; compound equality is now reference-based
(C# ReferenceEquals semantics).

Fix Neo N3 v3.10.0 consensus/protocol parity bugs (each verified against the
C# reference source):
- dBFT: gate Commit signing on holding every proposed transaction in
  check_prepare_responses, matching ConsensusService.CheckPreparations.
- Consensus: mint a fresh store snapshot each round so committee/validator/
  NextConsensus reads track the tip across committee-refresh heights, matching
  ConsensusContext.Reset (was reusing a frozen startup snapshot).
- StdLib.jsonSerialize: fault on invalid UTF-8 (StrictUTF8) instead of lossy
  U+FFFD substitution.
- CryptoLib.recoverSecp256K1: reject non-65-byte signatures, matching
  Crypto.ECRecover (a 64-byte EIP-2098 sig recovered here but returned null on
  C#). Adds a parity regression test.
- HSM: derive the System.Crypto.CheckSig interop hash from sha256(name)[..4]
  instead of a wrong hardcoded value.
- Oracle HTTPS: follow redirects only on 3xx; drop the spurious per-chunk 8 KB
  cap that could reject legitimate sub-limit responses.
- Wallet witness: validate signature length before the PUSHDATA1 length byte.

Three review findings were verified as C#-faithful / out-of-scope for v3.10.0
and deliberately left unchanged: the ArchivalNode 0x00 capability byte, the
mempool conflict fee accounting, and the HF_Gorgon GetCandidateVote override.

Also: converted compound-StackValue test assertions to structural comparison,
fixed a neo-system parallel test-isolation flake (shared poison-resilient
native-provider guard), and corrected pre-existing architecture-test drift
(neo-static-files classification, stale doc-reference paths).

All 3326 workspace tests pass; cargo fmt and clippy are clean.

+++ neo-vm/src/stack_item/array.rs
@@ -47,7 +47,7 @@ impl Array {
             inner: Arc::new(Mutex::new(ArrayInner {
                 items,
                 reference_counter,
-                id: next_stack_item_id(),
+                id: next_stack_item_id() as usize,
                 is_read_only: false,
             })),
         };
@@ -62,7 +62,7 @@ impl Array {
             inner: Arc::new(Mutex::new(ArrayInner {
                 items,
                 reference_counter: None,
-                id: next_stack_item_id(),
+                id: next_stack_item_id() as usize,
                 is_read_only: false,
             })),
         }
